Tast: Add support to cache extracted raw videos in encode tests.
As the raw video files used to test encoding are too large to download
and store on a DUT, videos are stored in the vp9.webm format. Before
each test these videos are extracted to create the raw input video.
This process can take a lot of time, so this CL adds support to cache
the extracted raw video file. This saves about 30s test time when
running 1080p encode testcases. The feature is disabled by default as
we don't want to have DUT's run out of disk space in the lab.

The option is added to the following Chrome video encode tests:
- video.DecodeEncodeAccelPerf.*
- video.EncodeAccel.*
- video.EncodeAccelPerf.*

BUG=None
TEST=various video encode tests on Hatch

+++ src/chromiumos/tast/local/bundles/cros/video/decode_encode_accel_perf.go
@@ -39,6 +39,8 @@ func init() {
 
 func DecodeEncodeAccelPerf(ctx context.Context, s *testing.State) {
 	const (
+		// Enable to cache the extracted raw video to speed up the test.
+		cacheExtractedVideo = false
 		// Time reserved for cleanup.
 		cleanupTime = 10 * time.Second
 		// Time to wait for CPU to stabilize after launching tests.
@@ -80,7 +82,9 @@ func DecodeEncodeAccelPerf(ctx context.Context, s *testing.State) {
 	if err != nil {
 		s.Fatal("Failed to prepare YUV file: ", err)
 	}
-	defer os.Remove(streamPath)
+	if !cacheExtractedVideo {
+		defer os.Remove(streamPath)
+	}
 
 	// Wait for the CPU to become idle.
 	if err := cpu.WaitUntilIdle(ctx); err != nil {
